move consts from LayerFactory to const.ts and parseSHInstanceId to utils.ts
