+++ example/node/index.js
@@ -156,7 +156,11 @@ async function run() {
   );
   printOut('flyovers for S1 GRD:', flyoversS1GRD);
 
-  const datesS1GRD = await layerS1.findDatesUTC(getMapParams.bbox, getMapParams.fromTime, getMapParams.toTime);
+  const datesS1GRD = await layerS1.findDatesUTC(
+    getMapParams.bbox,
+    getMapParams.fromTime,
+    getMapParams.toTime,
+  );
   printOut('dates for S1 GRD', datesS1GRD);
 
   // finally, display the image:
