+++ src/layer/utils.ts
@@ -3,6 +3,7 @@ import { stringify } from 'query-string';
 import { parseStringPromise } from 'xml2js';
 
 import { getAuthToken, isAuthTokenSet } from 'src/auth';
+import { SH_SERVICE_HOSTNAMES_V1_OR_V2, SH_SERVICE_HOSTNAMES_V3 } from './const';
 
 let fetchCache: Record<string, Promise<AxiosResponse>> = {};
 
@@ -76,8 +77,31 @@ export async function fetchGetCapabilitiesJson(baseUrl: string, forceFetch = fal
 }
 
 export async function fetchGetCapabilitiesJsonV1(baseUrl: string, forceFetch = false): Promise<any[]> {
-  const instanceId = this.parseSHInstanceId(baseUrl);
+  const instanceId = parseSHInstanceId(baseUrl);
   const url = `https://eocloud.sentinel-hub.com/v1/config/instance/instance.${instanceId}?scope=ALL`;
   const res = await fetchCached(url, { responseType: 'json' }, forceFetch);
   return res.data.layers;
+}
+
+export function parseSHInstanceId(baseUrl: string): string {
+  const INSTANCE_ID_LENGTH = 36;
+  // AWS:
+  for (let hostname of SH_SERVICE_HOSTNAMES_V3) {
+    const prefix = `${hostname}ogc/wms/`;
+    if (!baseUrl.startsWith(prefix)) {
+      continue;
+    }
+    const instanceId = baseUrl.substr(prefix.length, INSTANCE_ID_LENGTH);
+    return instanceId;
+  }
+  // EOCloud:
+  for (let hostname of SH_SERVICE_HOSTNAMES_V1_OR_V2) {
+    const prefix = `${hostname}v1/wms/`;
+    if (!baseUrl.startsWith(prefix)) {
+      continue;
+    }
+    const instanceId = baseUrl.substr(prefix.length, INSTANCE_ID_LENGTH);
+    return instanceId;
+  }
+  throw new Error(`Could not parse instanceId from URL: ${baseUrl}`);
 }
